refactor(types): замечания ревью — регистры вынесены из провайдера типов
`ConfigurationTypesProvider` перевалил порог Sonar в 750 строк кода (было 862).
Регистровая часть выделена в два класса по смыслу:

- `RegisterFamilies` — правило починки ссылки на чужое семейство регистров
  (ошибка синтакс-помощника), чистая функция от реестра;
- `RegisterTypesRegistrar` — то, что достраивается по конфигурации: expansion
  измерений/ресурсов/реквизитов, коллекция записей у набора и достройка членов
  с оставшимся плейсхолдером.

Провайдер стал 678 строк кода. Заодно устранены остальные замечания в этих
местах: висячий Javadoc (свой у `registerRecordExpansion`, TODO уехал в тело
метода) и два `continue` в одном цикле — тело цикла стало отдельным методом.
