+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/registry/ConfigurationTypesProvider.java
@@ -32,14 +32,10 @@ import com.github._1c_syntax.bsl.languageserver.infrastructure.WorkspaceContextH
 import com.github._1c_syntax.bsl.languageserver.infrastructure.WorkspaceScope;
 import com.github._1c_syntax.bsl.languageserver.types.model.BilingualString;
 import com.github._1c_syntax.bsl.languageserver.types.model.MemberDescriptor;
-import com.github._1c_syntax.bsl.languageserver.types.model.MemberKind;
 import com.github._1c_syntax.bsl.languageserver.types.model.MemberSource;
 import com.github._1c_syntax.bsl.languageserver.types.model.PlatformMetadata;
 import com.github._1c_syntax.bsl.languageserver.types.model.TypeRef;
-import com.github._1c_syntax.bsl.languageserver.types.model.SignatureDescriptor;
 import com.github._1c_syntax.bsl.languageserver.types.model.TypeSet;
-import com.github._1c_syntax.bsl.mdo.AccountingRegister;
-import com.github._1c_syntax.bsl.mdo.AccumulationRegister;
 import com.github._1c_syntax.bsl.mdo.Attribute;
 import com.github._1c_syntax.bsl.mdo.AttributeOwner;
 import com.github._1c_syntax.bsl.mdo.CalculationRegister;
@@ -48,7 +44,6 @@ import com.github._1c_syntax.bsl.mdo.ChartOfCalculationTypes;
 import com.github._1c_syntax.bsl.mdo.CommonAttribute;
 import com.github._1c_syntax.bsl.mdo.DocumentJournal;
 import com.github._1c_syntax.bsl.mdo.Enum;
-import com.github._1c_syntax.bsl.mdo.InformationRegister;
 import com.github._1c_syntax.bsl.mdo.MD;
 import com.github._1c_syntax.bsl.mdo.MDObject;
 import com.github._1c_syntax.bsl.mdo.PredefinedDataOwner;
@@ -70,7 +65,6 @@ import org.springframework.stereotype.Component;
 import java.util.ArrayList;
 import java.util.EnumSet;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -121,14 +115,6 @@ public class ConfigurationTypesProvider {
 
   /** Платформенный тип строки табличной части. */
   private static final String TABULAR_SECTION_ROW_TYPE = "Строка табличной части";
-
-  /**
-   * Ядра имён семейств регистров. Внутри семейства типы отличаются только суффиксом
-   * ({@code Запись}, {@code КлючЗаписи}, {@code НаборЗаписей} …), поэтому по этому списку
-   * опознаётся ссылка на чужое семейство (см. {@link #registerRegisterFamilyFixups}).
-   */
-  private static final Set<String> REGISTER_FAMILY_CORES = Set.of(
-    "РегистрСведений", "РегистрНакопления", "РегистрБухгалтерии", "РегистрРасчета");
 
   /** MDOType'ы, для которых имеет смысл регистрировать менеджер-тип. */
   private static final Set<MDOType> MANAGER_TYPES = Set.of(
@@ -191,6 +177,7 @@ public class ConfigurationTypesProvider {
   private final MetadataCollectionSpecializer metadataCollectionSpecializer;
   private final ConfigurationGenericExpander genericExpander;
   private final ServiceModuleEventRegistrar serviceModuleEventRegistrar;
+  private final RegisterTypesRegistrar registerTypesRegistrar;
   private final RecorderIndex recorderIndex;
   @Qualifier("platformTypesWarmupExecutor")
   private final AsyncTaskExecutor platformTypesWarmupExecutor;
@@ -366,7 +353,7 @@ public class ConfigurationTypesProvider {
     }
     var familyCore = fullName.getRu();
     registerFamilySpecializations(familyCore, name);
-    registerRegisterFamilyFixups(md, familyCore, name);
+    registerTypesRegistrar.registerFamilyFixups(md, familyCore, name);
     registerDerivedSpecializations(md, name);
     if (md instanceof DocumentJournal journal) {
       registerDocumentJournalColumnMembers(journal, familyCore, name);
@@ -377,10 +364,10 @@ public class ConfigurationTypesProvider {
     if (md instanceof PredefinedDataOwner predefinedDataOwner) {
       registerPredefinedValueExpansion(ref, familyCore, name, predefinedDataOwner);
     }
-    var registerChildren = registerChildrenOf(md);
+    var registerChildren = RegisterTypesRegistrar.registerChildrenOf(md);
     if (registerChildren != null) {
-      registerRecordSetCollectionMembers(familyCore, name);
-      registerRegisterRecordExpansion(familyCore, name, registerChildren);
+      registerTypesRegistrar.registerRecordSetCollectionMembers(familyCore, name);
+      registerTypesRegistrar.registerRecordExpansion(familyCore, name, registerChildren);
     }
   }
 
@@ -696,137 +683,6 @@ public class ConfigurationTypesProvider {
    * {@code СправочникСписок.<Имя>} и любые другие, которые HBK заведёт в
    * будущем.
    */
-  /**
-   * Достраивает типы членов регистра, которые специализация по имени регистра не
-   * покрывает. Оба случая видны по одному признаку — оставшемуся generic-плейсхолдеру:
-   * <ul>
-   *   <li>{@code Регистратор} объявлен как {@code ДокументСсылка.<Имя документа>}:
-   *       имя регистра к документу отношения не имеет, подставляются
-   *       документы-регистраторы (см. {@link RecorderIndex});</li>
-   *   <li>член ссылается на <b>чужое</b> семейство регистров: у
-   *       {@code РегистрБухгалтерииНаборЗаписей} метод {@code Вставить} объявлен
-   *       возвращающим {@code РегистрНакопленияЗапись.<Имя регистра накопления>}, хотя
-   *       соседние {@code Добавить} и {@code Получить} — {@code РегистрБухгалтерииЗапись}.
-   *       Это ошибка синтакс-помощника; семейство заменяется на своё.</li>
-   * </ul>
-   * Обе правки принимаются, только если получившийся тип есть в реестре.
-   */
-  // TODO mdclasses#670: так же типизировать Счет/СчетДт/СчетКт записи регистра
-  //  бухгалтерии (объявлены как ПланСчетовСсылка.<Имя плана счетов>) и ВидРасчета
-  //  записи регистра расчёта. План указан у регистра в метаданных (<ChartOfAccounts>,
-  //  <ChartOfCalculationTypes>), но mdclasses 0.20.0 его не разбирает. Как появится —
-  //  добавить имя плана вторым набором подстановки рядом с recorders: подходящий
-  //  выберется по тому же признаку «тип есть в реестре».
-  private void registerRegisterFamilyFixups(MD md, String familyCore, String mdName) {
-    if (!REGISTER_FAMILY_CORES.contains(familyCore)) {
-      return;
-    }
-    var recorders = recorderIndex.recordersOf(md.getMdoReference().getMdoRefRu());
-    for (var generic : typeRegistry.findAllGenericsByFamilyCore(familyCore)) {
-      var parameters = typeRegistry.getTypeParameters(generic);
-      if (parameters.size() != 1) {
-        continue;
-      }
-      var bindings = Map.of(parameters.get(0), mdName);
-      var specialized = typeRegistry.resolve(TypeRef.specialize(generic, bindings).qualifiedName())
-        .orElse(null);
-      if (specialized == null || specialized.equals(generic)) {
-        continue;
-      }
-      typeRegistry.registerMemberOverride(specialized,
-        () -> fixedUpRegisterMembers(generic, familyCore, bindings, mdName, recorders), FileType.BSL);
-    }
-  }
-
-  /**
-   * Члены generic-типа, у которых после подстановки имени регистра остался плейсхолдер,
-   * — с достроенными типами. Остальные не отдаются вовсе: они приходят обычной
-   * специализацией, и дублировать их незачем.
-   * <p>
-   * Читаются у generic'а, а не у специализированного типа: источник регистрируется как
-   * раз на специализированный, и обращение к его же {@code getMembers} зациклилось бы.
-   */
-  private List<MemberDescriptor> fixedUpRegisterMembers(TypeRef generic, String familyCore,
-                                                        Map<String, String> bindings, String mdName,
-                                                        List<String> recorders) {
-    var result = new ArrayList<MemberDescriptor>();
-    for (var member : typeRegistry.getMembers(generic, FileType.BSL)) {
-      if (member.generic()) {
-        continue;
-      }
-      var specialized = member.specialize(bindings);
-      var fixed = ownFamilyMember(specialized, familyCore, mdName);
-      if (fixed == null) {
-        fixed = PlaceholderBinder.bind(typeRegistry, specialized, recorders);
-      }
-      if (fixed != null) {
-        result.add(fixed);
-      }
-    }
-    return List.copyOf(result);
-  }
-
-  /**
-   * Заменяет в типах члена чужое семейство регистров на своё.
-   *
-   * @return член со своим типом; {@code null}, если подменять нечего.
-   */
-  private @Nullable MemberDescriptor ownFamilyMember(MemberDescriptor member, String familyCore,
-                                                     String mdName) {
-    var returnTypes = ownFamilyTypes(member.returnTypes(), familyCore, mdName);
-    if (returnTypes == null) {
-      return null;
-    }
-    var signatures = member.signatures().stream()
-      .map(signature -> ownFamilySignature(signature, familyCore, mdName))
-      .toList();
-    return member.withReturnTypes(returnTypes).withSignatures(signatures);
-  }
-
-  private SignatureDescriptor ownFamilySignature(SignatureDescriptor signature, String familyCore,
-                                                 String mdName) {
-    var returnTypes = ownFamilyTypes(signature.returnTypes(), familyCore, mdName);
-    if (returnTypes == null) {
-      return signature;
-    }
-    return new SignatureDescriptor(signature.parameters(), returnTypes,
-      signature.bilingualDescription(), signature.metadata());
-  }
-
-  /** @return набор с заменённым семейством; {@code null}, если ни один тип не поменялся. */
-  private @Nullable TypeSet ownFamilyTypes(TypeSet types, String familyCore, String mdName) {
-    var refs = new ArrayList<TypeRef>(types.refs().size());
-    var changed = false;
-    for (var ref : types.refs()) {
-      var own = ownFamilyRef(ref, familyCore, mdName);
-      changed |= !own.equals(ref);
-      refs.add(own);
-    }
-    return changed ? TypeSet.of(refs) : null;
-  }
-
-  private TypeRef ownFamilyRef(TypeRef ref, String familyCore, String mdName) {
-    var qualifiedName = ref.qualifiedName();
-    if (ref.placeholders().isEmpty() || qualifiedName.startsWith(familyCore)) {
-      return ref;
-    }
-    var dot = qualifiedName.indexOf('.');
-    if (dot < 0) {
-      return ref;
-    }
-    for (var foreignCore : REGISTER_FAMILY_CORES) {
-      if (!qualifiedName.startsWith(foreignCore)) {
-        continue;
-      }
-      var kindSuffix = qualifiedName.substring(foreignCore.length(), dot);
-      var own = typeRegistry.resolve(familyCore + kindSuffix + "." + mdName).orElse(null);
-      if (own != null) {
-        return own;
-      }
-    }
-    return ref;
-  }
-
   private void registerFamilySpecializations(String familyCore, String mdName) {
     for (var generic : typeRegistry.findAllGenericsByFamilyCore(familyCore)) {
       var parameters = typeRegistry.getTypeParameters(generic);
@@ -928,131 +784,6 @@ public class ConfigurationTypesProvider {
   }
 
   /**
-   * Триплет имён детей регистра (измерения/ресурсы/реквизиты), полученный
-   * из конкретного MD-класса регистра. {@code null} — для не-регистров.
-   */
-  record RegisterChildren(List<? extends Attribute> dimensions,
-                          List<? extends Attribute> resources,
-                          List<? extends Attribute> attributes) {
-  }
-
-  static @Nullable RegisterChildren registerChildrenOf(MD md) {
-    return switch (md) {
-      case InformationRegister r ->
-        new RegisterChildren(r.getDimensions(), r.getResources(), customAttributesOf(r.getAttributes()));
-      case AccumulationRegister r ->
-        new RegisterChildren(r.getDimensions(), r.getResources(), customAttributesOf(r.getAttributes()));
-      case AccountingRegister r ->
-        new RegisterChildren(r.getDimensions(), r.getResources(), customAttributesOf(r.getAttributes()));
-      case CalculationRegister r ->
-        new RegisterChildren(r.getDimensions(), r.getResources(), customAttributesOf(r.getAttributes()));
-      default -> null;
-    };
-  }
-
-  /**
-   * Отфильтровывает {@link StandardAttribute} (Период/Регистратор/Активность/…):
-   * {@code getAttributes()} регистра возвращает их вперемешку с собственными
-   * реквизитами, но они уже приходят как обычные bilingual-члены generic-типа
-   * записи ({@code РегистрХХХЗапись.<Имя>}) из bsl-context — без фильтра
-   * плейсхолдер {@code <Имя реквизита>} материализовал бы их второй раз,
-   * одноязычными (под английским написанием).
-   */
-  private static List<? extends Attribute> customAttributesOf(List<? extends Attribute> attributes) {
-    return attributes.stream().filter(a -> !(a instanceof StandardAttribute)).toList();
-  }
-
-  /**
-   * Регистрирует expansion generic-property {@code <Имя измерения>/<Имя ресурса>/
-   * <Имя реквизита>} на типе записи регистра ({@code РегистрСведенийЗапись.<Имя>}
-   * и аналоги для других семейств регистров). Имена детей берутся из mdclasses;
-   * мета — наследуется от HBK-template'ов.
-   *
-   * @param familyCore ru-часть имени семейства ({@code "РегистрСведений"} и т.п.)
-   * @param regName    имя регистра в конфигурации
-   * @param children   измерения/ресурсы/реквизиты регистра
-   */
-  /**
-   * Делает набор записей регистра полноценной коллекцией своих записей: обход
-   * {@code Для Каждого} и индексатор дают запись <b>этого</b> регистра, а
-   * {@code Выгрузить()} — таблицу значений с его колонками.
-   * <p>
-   * Платформа описывает набор обобщённо: {@code Выгрузить()} объявлен возвращающим
-   * просто {@code ТаблицаЗначений}, хотя колонки у выгруженной таблицы те же, что у
-   * записи. Приём тот же, что у табличных частей (см.
-   * {@link #registerTabularSectionPlatformMembers}); колонками служат свойства записи —
-   * измерения, ресурсы, реквизиты и стандартные реквизиты вроде {@code Регистратор}.
-   *
-   * @param familyCore ru-часть имени семейства ({@code "РегистрНакопления"} и т.п.).
-   * @param regName    имя регистра в конфигурации.
-   */
-  private void registerRecordSetCollectionMembers(String familyCore, String regName) {
-    var genericSet = singleGenericOf(familyCore + "НаборЗаписей");
-    var genericRecord = singleGenericOf(familyCore + "Запись");
-    if (genericSet == null || genericRecord == null) {
-      return;
-    }
-    var setRef = specializedByName(genericSet, regName);
-    var recordRef = specializedByName(genericRecord, regName);
-    if (setRef == null || recordRef == null) {
-      return;
-    }
-    // Тип элемента — до наследования коллекционных свойств: иначе выиграла бы
-    // унаследованная обобщённая запись и обход потерял бы колонки.
-    typeRegistry.registerDefaultElementTypes(setRef, List.of(recordRef));
-    typeRegistry.inheritCollectionTraits(setRef, genericSet, FileType.BSL);
-
-    var valueTableRow = typeRegistry.resolve(CollectionReturnsSpecializer.VALUE_TABLE_ROW).orElse(null);
-    MemberSource columns = () -> typeRegistry.getMembers(recordRef, FileType.BSL).stream()
-      .filter(member -> member.kind() == MemberKind.PROPERTY && !member.generic())
-      .toList();
-    typeRegistry.registerMemberOverride(setRef, () -> CollectionReturnsSpecializer.specialize(
-      typeRegistry.getMembers(genericSet, FileType.BSL), genericRecord, recordRef,
-      CollectionReturnsSpecializer.unloadedRow(valueTableRow, columns)), FileType.BSL);
-  }
-
-  /** Единственный generic семейства с ровно одним параметром; {@code null} — такого нет. */
-  private @Nullable TypeRef singleGenericOf(String familyCore) {
-    return typeRegistry.findAllGenericsByFamilyCore(familyCore).stream()
-      .filter(generic -> typeRegistry.getTypeParameters(generic).size() == 1)
-      .findFirst()
-      .orElse(null);
-  }
-
-  /** Специализация generic'а по имени MD-объекта; {@code null} — не зарегистрирована. */
-  private @Nullable TypeRef specializedByName(TypeRef generic, String mdName) {
-    var bindings = Map.of(typeRegistry.getTypeParameters(generic).get(0), mdName);
-    return typeRegistry.resolve(TypeRef.specialize(generic, bindings).qualifiedName()).orElse(null);
-  }
-
-  private void registerRegisterRecordExpansion(String familyCore, String regName, RegisterChildren children) {
-    var generic = typeRegistry.findAllGenericsByFamilyCore(familyCore + "Запись").stream()
-      .findFirst()
-      .orElse(null);
-    if (generic == null) {
-      return;
-    }
-    var parameters = typeRegistry.getTypeParameters(generic);
-    if (parameters.size() != 1) {
-      return;
-    }
-    var typeBindings = Map.of(parameters.get(0), regName);
-    var specializedName = TypeRef.specialize(generic, typeBindings).qualifiedName();
-    var specialized = typeRegistry.resolve(specializedName).orElse(null);
-    if (specialized == null) {
-      return;
-    }
-    var expansions = new LinkedHashMap<String, List<String>>();
-    putAttributeNames(expansions, "Имя измерения", children.dimensions());
-    putAttributeNames(expansions, "Имя ресурса", children.resources());
-    putAttributeNames(expansions, "Имя реквизита", children.attributes());
-    if (expansions.isEmpty()) {
-      return;
-    }
-    typeRegistry.registerMemberExpansion(specialized, generic, typeBindings, expansions, FileType.BSL);
-  }
-
-  /**
    * Графы журнала документов как property-члены на типе журнала
    * ({@code ЖурналДокументов.<имя>}). Источник имён и типов — mdclasses
    * ({@link DocumentJournal#getColumns()}); все колонки реализуют
@@ -1096,18 +827,6 @@ public class ConfigurationTypesProvider {
       if (!recalcName.isBlank()) {
         registerFamilySpecializations("Перерасчет", recalcName);
       }
-    }
-  }
-
-  /** Кладёт в expansion-map имена непустых атрибутов под ключом-placeholder'ом. */
-  static void putAttributeNames(Map<String, List<String>> sink, String placeholder,
-                                List<? extends Attribute> attributes) {
-    var names = attributes.stream()
-      .map(Attribute::getName)
-      .filter(n -> !n.isBlank())
-      .toList();
-    if (!names.isEmpty()) {
-      sink.put(placeholder, names);
     }
   }
